Made a bunch of experimental changes for a animation workflow
Animation tools: Added a whole section below where you import animations:
Create IK bones for both arms and legs or individually
Allign those ik bones to the original bones locations for just the frame you are on or the ENTIRE animation
Apply IK animation: basically baking the action but gets rid of the ik bones so you dont have to remove em manually
Toggle IK Influence: with this you can set how much you want the ik bones to affect the og bones for a specific frame, this is good for mixing IK bones with the og bones to get some interesting results, I suggest experimenting with it
Position IK to Match FK: Havent tried it ngl, I made that one and never continued working on it so I doubt it works
Transfer Hip Animation to Trans: this is for retargetted animations where the character moves forward but the trans bone stays in place. With it the forward movement gets transferred to the trans bone
Reset Bone Locations: this one is very cool, lets say you have the animation of ganondorf on a smaller character, with this button it basically makes the animation fit the new character so it doesnt look either too stretched or too compressed.
Idle Pose Library: Lets say your animation ends on frame 21 but it doesnt go back to idle which is necessary, just go to frame 25 and press the button and it will put the first frame of a00wait1 there if its loaded, basically saving you the hassle of having to load the anim and copy the first frame.
Invert Positive And Negative: Inverts values of the current keyframe of the selected bone.

Model Importer: Now you can browse your mod directory and it will show all your mods on a list to easily import the base model, if you want to import a different alt, there is also a button to browse for individual model and import it.

Animation Importer: This one is really cool, lets say you imported Broly for example, the plugin will search for a motion folder in the folder where Broly is and if it finds it it will populate the entire animation section with all of the anims for you to easily import the one you want, you can also choose a different Animation folder or a specific animation if you want. There is also a feature to import ALL the animation on the list, this is more catered for bulk anim works like bulk retargets.

Animation Exporter: Added support to export all the animations loaded on the blend file, great for bulk anim work.

Smart un-exo model: This tool lets you grab a exo model and remove the exo from it in a smart and clean way.
Cleanup Unused Exo Bones: This button removes all the bones that arent related to the smash rig. (Select rig and click)
Transfer Exo Weights: For this one select ALL meshes and the armature to use it, its transfers the weights from the exo to the smash rig.
Align Smash Bones to Exo Bones: select armature and go into EDIT MODE and press the button, it will do what the name suggests. After that just remove the exo bones and position smash bones correctly.

Swing: Now you can auto detect all swing bone chains, copy/paste swing values, make presets for a single bone or a swing bone chain. I also fixed an issue where if you made custom names for your chains, the next time you loaded the swing.prc they would be hashed, now they properly save on the ParamLabels.csv so when you import it again, the names will load properly.

Smash Material Conversion: When you are making a model, you can convert your Principled BSDF material to a smash material, that was there already but now you can bake a prm map based on the values you modify on the Principled BSDF or if you have textures like metalness, roughess, etc.